Migrate new cogs from cogs2/ to cogs/ folder, effectively deleting all d.py files
