+++ cogs/images.py
