+++ cogs/mcserv.py
@@ -1,48 +1,42 @@
-import discord
+from mcstatus import JavaServer
+from utils.embed import create_embed
+from variables import server
+import interactions
 import socket
-import variables
-from discord.ext import commands
-from mcstatus import MinecraftServer
 
 
-class MCServ(commands.Cog):
-    def __init__(self, client):
-        self.client = client
-        self.servers = [
-            {
-                "ip": "btefr.thesmyler.fr:7000",
-                "desc": "Serveur Moddé 1.12 - IP: buildtheearth.net, /bt FR",
-            },
-        ]
-        self.mc_embed = discord.Embed(
-            title="**Statut du serveur Minecraft BTE France**", description="_ _\n"
-        )
-        self.mc_embed.set_thumbnail(url=variables.bte_france_icon)
+# List of servers to get the status from.
+servers = [
+    {
+        "ip": "btefr.thesmyler.fr:7000",
+        "desc": "Serveur Moddé 1.12 - IP: buildtheearth.net, /bt FR",
+    }
+]
 
-    @commands.command(
-        brief="Statut du serveur Minecraft", aliases=["mcserv", "mcstatus", "status"]
-    )
-    async def mc(self, ctx):
-        self.mc_embed.clear_fields()
-        for server in self.servers:
+
+class MCServ(interactions.Extension):
+    def __init__(self, client: interactions.Client):
+        self.client: interactions.Client = client
+
+    @interactions.extension_command(name="mc", description="Statut du serveur Minecraft", scope=server)
+    async def mc(self, ctx: interactions.CommandContext):
+        fields = []
+        for server_dict in servers:
+
             try:
-                status = MinecraftServer.lookup(server["ip"]).status()
+                status = JavaServer.lookup(server_dict["ip"]).status()
+
             except (ConnectionRefusedError, socket.timeout):
                 embed_value = ":x: Serveur hors ligne!"
-                embed_value += "\n\n_ _" if server == self.servers[0] else ""
-                self.mc_embed.add_field(
-                    name=server["desc"], value=embed_value, inline=False
-                )
+                embed_value += "\n\n_ _" if server_dict == servers[0] else ""
+                fields.append(interactions.EmbedField(name=server_dict["desc"], value=embed_value, inline=False))
+
             else:
                 online_players = status.players.online
-                sample = (
-                    sorted([p.name for p in status.players.sample])
-                    if status.players.sample is not None
-                    else []
-                )
-                sample_title = (
-                    "Joueurs Connectés" if online_players != 1 else "Joueur Connecté"
-                )
+                sample = sorted([p.name for p in status.players.sample]) if status.players.sample is not None else []
+
+                sample_title = "Joueurs Connectés" if online_players != 1 else "Joueur Connecté"
+
                 sample_txt = ""
                 if online_players != 0:
                     diff = online_players - len(sample)
@@ -53,16 +47,17 @@ class MCServ(commands.Cog):
                         sample_txt = sample[0]
                     else:
                         sample_txt = ", ".join(sample[:-1]) + " et " + sample[-1]
-                embed_value = (
-                    f":white_check_mark: Serveur en ligne!\n\n**{online_players} {sample_title}**\n"
-                    + discord.utils.escape_markdown(sample_txt)
-                )
-                # embed_value += '\n\n_ _' if server == self.servers[0] else ''
-                self.mc_embed.add_field(
-                    name=server["desc"], value=embed_value, inline=False
-                )
-        await ctx.send(embed=self.mc_embed)
+                sample_txt = sample_txt.replace("_", r"\_").replace("*", r"\*")  # Escape markdown symbols that could be present in players' names
+
+                embed_value = f":white_check_mark: Serveur en ligne!\n\n**{online_players} {sample_title}**\n" + sample_txt
+                fields.append(interactions.EmbedField(name=server_dict["desc"], value=embed_value, inline=False))
+
+        await ctx.send(embeds=create_embed(
+            title="**Statut du serveur Minecraft BTE France**",
+            description="_ _\n",
+            fields=fields
+        ))
 
 
-def setup(client):
-    client.add_cog(MCServ(client))
+def setup(client: interactions.Client):
+    MCServ(client)
