+++ cogs/warps.py
@@ -1,80 +1,119 @@
-import discord
-import variables
-from discord.ext import commands
+from utils.embed import create_embed
+from variables import server
+import interactions
+import random
 
-warps = {
-    "La Défense": [
+
+class Warp:
+    def __init__(self, name, warp, description, image):
+        self.name = name
+        self.warp = warp
+        self.description = description
+        self.image = image
+
+
+WARPS = [
+    Warp(
+        "La Défense",
         "defense_esplanade",
         "Quartier d'affaires de Paris connu pour ses nombreux gratte-ciels.\nBusiness district of Paris known for its numerous skyscrapers.",
-        "https://media.discordapp.net/attachments/694325296106569748/719921092814438460/2020-06-09_10.27.30.png?width=1194&height=671",
-    ],
-    "Arc de Triomphe": [
-        "arc_de_triomphe",
+        "https://imgur.com/ZOTj9dn.png"
+    ),
+    Warp(
+        "Arc de Triomphe",
+        "paris:etoile",
         "Arc au coeur de Paris, situé sur l'avenue des Champs-Elysées.\nArc at the heart of Paris, situated on the Avenue des Champs-Elysées.",
-        "https://images-ext-1.discordapp.net/external/zzWs72V7Gglq3vHYKV9r33NAhYssRQ9WhDFeL0XMqZM/https/i.imgur.com/ey0Dewp.png?width=641&height=672",
-    ],
-    "Île Saint-Louis": [
-        "saint-louis",
-        "Île en plein milieu de Paris et bordée par la Seine.\nIsland in the middle of Paris and bordered by the Seine.",
-        "https://media.discordapp.net/attachments/694325296106569748/719704689318559795/2020-06-09_02.00.20.png?width=1268&height=671",
-    ],
-    "Nantes": [
-        "nantes",
+        "https://imgur.com/OBemDc5.png"
+    ),
+    Warp(
+        "Ile de la Cité",
+        "Ile_de_la_Cite",
+        "Le berceau de la ville de Paris, construit par la team ETB.\nThe cradle of Paris, built by the ETB team.",
+        "https://imgur.com/FcLGAae.png"
+    ),
+    Warp(
+        "Nantes",
+        "Nantes:cathedrale",
         "Cité des Ducs de Bretagne, non loin de l'océan Atlantique.\nCity of Dukes of Brittany, not far from the Atlantic ocean.",
-        "https://i.imgur.com/DW9k26J.png",
-    ],
-    "Turenne": [
+        "https://imgur.com/WM4frCq.png"
+    ),
+    Warp(
+        "Turenne",
         "turenne",
         "Village en Corrèze, réalisé en 48h en tant qu'évènement communautaire.\n Village in Corrèze, realized in 48h as a community event.",
-        "https://images-ext-1.discordapp.net/external/pbkeBOTWQVA-IUa8blek5pNI-sUZhNIPu_Yw-WgLM-U/https/i.imgur.com/hvd2klJ.jpg?width=1194&height=671",
-    ],
-    "Belle-Île-en-Mer": [
-        "belle_ile",
-        "Île dans l'océan Atlantique ayant été une forteresse auparavant.\nIsland in the Atlantic ocean having been previously a fortress.",
-        "https://media.discordapp.net/attachments/694325296106569748/708468079658795018/2020-05-09_01.34.14.png?width=1268&height=671",
-    ],
-    "Quimper": [
-        "quimper",
-        "Chef-lieu du Finistère, le centre ayant été construit en 96h en tant qu'évènement communautaire.\nCapital of Finistère, the center having been built in 48h as a community event.",
-        "https://media.discordapp.net/attachments/692259936012468235/714577660742729738/Image_04_000.png?width=1194&height=671",
-    ],
-    "Bordeaux": [
-        "bordeaux",
-        "Cinquième ville de France, connue mondialement pour son vin.\nFifth city of France, known worldwide for its wine.",
-        "https://media.discordapp.net/attachments/694325296106569748/710989490688360469/Image_01_005.png?width=1443&height=641",
-    ],
-    "Le Puy-Notre-Dame": [
+        "https://imgur.com/j5qThvh.png"
+    ),
+    Warp(
+        "Le Puy-Notre-Dame",
         "lepuynotredame",
         "Petite commune dans le Maine-et-Loire.\nSmall town in the Maine-et-Loire.",
-        "https://images-ext-1.discordapp.net/external/VcCWLicW7-sua-rBPR9cwcy4D2GYGAAW1tSWxOf1-B8/https/i.imgur.com/aG42ecr.png?width=1283&height=671",
-    ],
-    "Pointe Saint-Mathieu": [
-        "pointe_saint-mathieu",
-        "Pointe du Finistère, au bord des falaises.\nHeadland of Finistère, flanked by cliffs.",
-        "https://media.discordapp.net/attachments/694325296106569748/733279570395922462/2020-07-16_13.08.23.png?width=1268&height=671",
-    ],
-}
+        "https://imgur.com/VFyfKzR.png"
+    ),
+    Warp(
+        "Le Mont-Saint-Michel",
+        "mont-saint-michel",
+        "Îlot roché situé en Normandie, une des icones de la France.\nTidal island located in Normandy, one of the icons of France.",
+        "https://imgur.com/M9nsIbr.png"
+    ),
+    Warp(
+        "Turckheim",
+        "Turckheim",
+        "Petite commune en Alsace avec notamment beaucoup de maisons à colombages.\nSmall town in Alsace with many half-timbered houses.",
+        "https://imgur.com/qsjFdo6.png"
+    ),
+    Warp(
+        "Monaco",
+        "monaco",
+        "Principauté située sur la côte méditéranéenne, construite par la team Build It 1:1.\nPrincipality located on the Mediterranean coast, built by the Build It 1:1 team.",
+        "https://imgur.com/gA9DGCf.png"
+    ),
+    Warp(
+        "Viaduc de Millau",
+        "viaduc_de_millau",
+        "Le viaduc avec les pylônes les plus hauts du monde (343m).\nThe viaduct with the highest pylons in the world (343m).",
+        "https://imgur.com/P7ex8lG.png"
+    )
+]
 
 
-class Warps(commands.Cog):
-    def __init__(self, client):
-        self.client = client
-        self.thumbnail_url = variables.bte_france_icon
+class Warps(interactions.Extension):
+    def __init__(self, client: interactions.Client):
+        self.client: interactions.Client = client
 
-    @commands.command(brief="Liste des meilleurs Warps", aliases=["warp"])
-    async def warps(self, ctx):
-        if ctx.channel.type != discord.ChannelType.private:
-            await ctx.send(f"{ctx.author.mention}, regarde tes MPs! :mailbox:")
-        await ctx.author.send(
-            ":flag_fr: **Liste des meilleurs warps sur le serveur BTE France** :flag_fr:"
-        )
-        for name, desc in warps.items():
-            embed = discord.Embed(title=name, colour=discord.Colour(0xFFFFFF))
-            embed.add_field(name=f"/warp {desc[0]}", value=desc[1], inline=False)
-            embed.set_image(url=desc[2])
-            embed.set_thumbnail(url=self.thumbnail_url)
-            await ctx.author.send(embed=embed)
+    @interactions.extension_command(name="warps", description="List of the best BTE France warps", scope=server)
+    async def warps(self, ctx: interactions.CommandContext):
+        await ctx.send("Regarde tes MPs! :mailbox:")
+
+        # Randomize warps and chunk in 10 warps (max embeds per message is 10)
+        random.shuffle(WARPS)
+        chunked_warps_list = [WARPS[i:i + 10] for i in range(0, len(WARPS), 10)]
+
+        # Create all embeds
+        chunked_embeds = []
+        for warp_list in chunked_warps_list:
+            embed_list = [
+                create_embed(
+                    title=warp.name,
+                    color=0xFFFFFF,
+                    fields=[
+                        interactions.EmbedField(
+                            name=f"/warp {warp.warp}",
+                            value=warp.description,
+                            inline=False
+                        )
+                    ],
+                    image=warp.image
+                )
+                for warp in warp_list
+            ]
+            chunked_embeds.append(embed_list)
+
+        for i, embed_list in enumerate(chunked_embeds):
+            await ctx.author.send(
+                ":flag_fr: **Liste des meilleurs warps sur le serveur BTE France** :flag_fr:" if i == 0 else "",
+                embeds=embed_list
+            )
 
 
-def setup(client):
-    client.add_cog(Warps(client))
+def setup(client: interactions.Client):
+    Warps(client)
